+++ src/components/App.svelte
@@ -1,0 +1,15 @@
+<script>
+  import AppBar from './AppBar.svelte';
+  import WebView from './WebView.svelte';
+</script>
+
+<main>
+  <AppBar />
+  <WebView />
+</main>
+
+<style>
+  main {
+    height: 100%;
+  }
+</style>