+++ src/components/AppBar.svelte
@@ -1,0 +1,9 @@
+<script>
+  import ClientBar from './ClientBar.svelte';
+  import SearchBar from './SearchBar.svelte';
+</script>
+
+<div class="app-bar">
+  <ClientBar />
+  <SearchBar />
+</div>