+++ src/components/ClientBar.svelte
@@ -1,0 +1,20 @@
+<div class="client-bar">
+  <nav class="client-m drag">
+    <div class="client-m__btns--left">
+      <span class="client-m__btn click pin-btn">
+        <i class="fa fa-unlock js-pin"></i>
+      </span>
+    </div>
+    <div class="client-m__btns--right">
+      <span class="client-m__btn click minimize-btn">
+        <i class="fa fa-window-minimize"></i>
+      </span>
+      <span class="client-m__btn click maximize-btn">
+        <i class="fa fa-window-maximize"></i>
+      </span>
+      <span class="client-m__btn click close-btn">
+        <i class="fa fa-times"></i>
+      </span>
+    </div>
+  </nav>
+</div>