+++ src/components/SearchBar.svelte
@@ -1,0 +1,22 @@
+<form class="search-bar">
+  <nav class="search-bar__nav">
+    <span class="client-m__btn click prev-btn inactive prev-btn">
+      <i class="fa fa-arrow-left"></i>
+    </span>
+    <span class="client-m__btn click refresh-btn inactive home-btn">
+      <i class="fa fa-home"></i>
+    </span>
+    <span class="client-m__btn click next-btn inactive next-btn">
+      <i class="fa fa-arrow-right"></i>
+    </span>
+  </nav>
+  <input
+    id="search-input"
+    class="search-bar__item"
+    type="text"
+    placeholder="Url here..."
+  />
+  <button class="search-bar__item search-submit-btn" type="submit">
+    Go
+  </button>
+</form>