+++ src/components/WebView.svelte
@@ -1,0 +1,3 @@
+<div class="webview">
+  <webview src="https://m.youtube.com"></webview>
+</div>