+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## [1.?.?] - ???
+
+- Feat: Svelte rollup build setup.
+- Feat: html moved to svelte components.
+
+---
+
 ## [1.4.1] - 2020-06-11
 
 ### Changed